backend/requirements.txt
pandas
numpy
sentence-transformers
scikit-learn
pdfplumber
python-docx
openpyxl
fastapi
uvicorn
openai
faiss-cpu
python-dotenv
torch
